refactor: utils file to helper
