+++ src/components/button/ButtonLink.tsx
@@ -4,7 +4,7 @@ import { Link, LinkProps } from "@/components/Link"
 import { ButtonVariants } from "@/components/button/variants"
 import type { VariantProps } from "class-variance-authority"
 import React from "react"
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/helper"
 
 interface ButtonLinkProps
   extends LinkProps,
