+++ src/components/ui/Badge.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/helper"
 import { cva, VariantProps } from "class-variance-authority"
 
 // The badge system:
