+++ src/components/ui/Card.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/helper"
 
 const Card = React.forwardRef<
   HTMLDivElement,
