+++ src/components/ui/Dialog.tsx
@@ -2,7 +2,7 @@
 
 import * as React from "react"
 import * as DialogPrimitive from "@radix-ui/react-dialog"
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/helper"
 import { IconButton } from "@/components/button/IconButton"
 
 const Dialog = DialogPrimitive.Root
